requirements.txt
arabic-reshaper==2.1.3
asgiref==3.4.1
astroid==2.4.2
backports.entry-points-selectable==1.1.0
certifi==2019.11.28
chardet==3.0.4
colorama==0.4.3
defusedxml==0.6.0
diff-match-patch==20200713
distlib==0.3.2
dj-static==0.0.6
Django==3.2.6
django-allauth==0.42.0
django-ckeditor==6.1.0
django-crispy-forms==1.9.2
django-el-pagination==3.3.0
django-hashers-passlib==0.1
django-import-export==2.5.0
django-js-asset==1.2.2
django-nose==1.4.7
django-rename-app==0.1.2
et-xmlfile==1.1.0
filelock==3.0.12
future==0.18.2
html5lib==1.1
idna==2.8
isort==5.5.4
lazy-object-proxy==1.4.3
MarkupPy==1.14
mccabe==0.6.1
nose==1.3.7
oauthlib==3.1.0
odfpy==1.4.1
openpyxl==3.0.7
passlib==1.7.4
Pillow==8.2.0
pipenv==2021.5.29
platformdirs==2.2.0
pylint==2.6.0
PyPDF2==1.26.0
python-bidi==0.4.2
python3-openid==3.2.0
pytz==2021.1
PyYAML==5.4.1
reportlab==3.5.68
requests==2.24.0
requests-oauthlib==1.3.0
six==1.13.0
sqlparse==0.4.1
sslcommerz-lib==1.0
static3==0.7.0
stripe==2.58.0
tablib==3.0.0
toml==0.10.1
urllib3==1.25.7
virtualenv==20.7.2
virtualenv-clone==0.5.6
virtualenvwrapper-win==1.2.6
webencodings==0.5.1
wrapt==1.11.2
xhtml2pdf==0.2.5
xlrd==2.0.1
xlwt==1.3.0